Inline everything into the HTML
I have noticed that the CSS, font and favicons are inherited from the
proxied site. This leads to some problems with the styling. I do not
want to rely on another sites assets. For this reason, we will forego
all local asset loading in favor of inlining:
- The CSS was inlined into a `style` tag
- The font was inlined into the CSS `@font-face`
- The favicon was inlined into the `link` of the `head`

Since we no longer have CSS file, the `lint-css` script was removed,
along with the stylelint dependency.